A lot of work regarding the introduction of torrent streaming to the project - the worst part is how different browsers implement the video tag and stream video. The only ones that seem to correctly accept partial byte-range responses are Chrome and Chromecast. videojs doesn't seem to alleviate this problem in any way. Will update the readme with this.
Added torrent directive which is not currently being used, may or may not need it. Using YIFY torrent API to fetch torrent information although that service has been down the last few days, which is why the message, "Something went wrong - we couldn't retrieve torrent information for this movie" will appear.

Leaving all this stuff in the dev branch until i've cleaned it up as much as possible.